test: Extract integration test debug setup to DebugIntegrationLogsMixin
